Add RBAC functionality (#2)
Roles and Permissions are now customizable in the PropelAuth backend, and this new change handles that. A user still has a single role, but now we also include a list of all roles that role "inherits" from, plus a list of permissions associated with that role. These are populated in OrgMemberInfo.

Changes to be aware of:

- OrgMemberInfo is now a class, and comes with four validators. All OrgMemberInfo objects are automatically created from the incoming JWT token.
- Those four validators have appropriate wrappers. These are what are ultimately exposed to the end user via initBaseAuth().
- Removed the concept of minimum required role and valid role. The new validators handles that functionality.

To painlessly upgrade your old code:

- If you want to validate a token, but you don't need to validate access, use validateAccessTokenAndGetUserWithOrgInfo().
- If you want to validate a token and you also want to validate access (in the same way it used to work), use validateAccessTokenAndGetUserWithOrgInfoWithMinimumRole().